Code clean up in NetLib: 1. Add GLOBAL_REMOVE_IF_UNREFERENCED to all globals 2. Update NTOHL and NTOHS to be BaseLib func SwapBytes32/SwapBytes16 3. Remove duplicate NET_SWAP_SHORT (to use NTOHS instead)
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@9648 6f19259b-4bc3-4df7-8a09-765794883524

+++ MdeModulePkg/Universal/Network/IScsiDxe/IScsiProto.c
@@ -1857,7 +1857,7 @@ IScsiNewScsiCmdPdu (
   CopyMem (ScsiCmd->Cdb, Packet->Cdb, sizeof (ScsiCmd->Cdb));
 
   if (Packet->CdbLength > 16) {
-    Header->Length  = NTOHS (Packet->CdbLength - 15);
+    Header->Length  = NTOHS ((UINT16) (Packet->CdbLength - 15));
     Header->Type    = ISCSI_AHS_TYPE_EXT_CDB;
 
     CopyMem (Header + 1, (UINT8 *) Packet->Cdb + 16, Packet->CdbLength - 16);
